requirements.txt
appdirs==1.4.3
attrs==16.3.0
Automat==0.5.0
cffi==1.9.1
constantly==15.1.0
cryptography==1.7.2
cssselect==1.0.1
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.6
django-toolbelt==0.0.1
enum34==1.1.6
gunicorn==19.7.0
idna==2.5
incremental==16.10.1
ipaddress==1.0.18
lxml==3.7.3
olefile==0.44
packaging==16.8
parsel==1.1.0
Pillow==4.0.0
psycopg2==2.6.2
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==16.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
queuelib==1.4.2
Scrapy==1.3.2
scrapy-djangoitem==1.1.1
service-identity==16.0.0
six==1.10.0
static3==0.7.0
Twisted==17.1.0
w3lib==1.17.0
zope.interface==4.3.3
